- Removed AjaxController - Moved methods from ajax controller to related controllers - Added comments to models and relationships - Updated web.php
