+++ app/Models/SalesPrice.php
@@ -13,6 +13,11 @@ class SalesPrice extends Model
     protected $table = "sales_prices";
     public $primaryKey = "id";
 
+    /**
+     * Returns an object of product.
+     *
+     * @return BelongsTo
+     */
     public function product(): BelongsTo
     {
         return $this->belongsTo(Product::class, 'product_id', 'id');
